Checking AI Line Of Sight
If the AI Pawn doesn't find a target or if the player pawn hides from AI Pawn's sight the focus of AI Pawn will be cleared and it should stop moving.
In the if block _______
if AI Pawn finds Player Pawn in it's line of sight then It will follow Player Pawn and make focus to Player Pawn.